+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "app.kushagra.takeoff"
+        applicationId "app.kushagra.appledental"
         minSdkVersion 16
         targetSdkVersion 29
         versionCode 1
