+++ app/src/main/java/app/kushagra/appledental/base/CoordinatedFragment.kt
@@ -1,0 +1,8 @@
+package app.kushagra.appledental.base
+
+import androidx.fragment.app.Fragment
+
+abstract class CoordinatedFragment : Fragment() {
+
+    abstract fun getTitle():String
+}
