+++ app/src/main/java/app/kushagra/appledental/extensions/EasyExtension.kt
@@ -1,0 +1,2 @@
+package app.kushagra.appledental.extensions
+
