+++ app/src/main/java/app/kushagra/appledental/model/QuickDataModel.kt
@@ -1,0 +1,10 @@
+package app.kushagra.appledental.model
+
+import android.graphics.drawable.Drawable
+
+data class QuickDataModel(
+    var title:String?="",
+    var image:Drawable?=null,
+    var background:Int? = null,
+    var mapId:Int?=0
+)
