+++ app/src/main/java/app/kushagra/appledental/sections/MainActivity.kt
@@ -1,0 +1,12 @@
+package app.kushagra.appledental.sections
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import app.kushagra.appledental.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
