+++ app/src/main/java/app/kushagra/appledental/sections/doctor/DoctorActivity.kt
@@ -1,0 +1,12 @@
+package app.kushagra.appledental.sections.doctor
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import app.kushagra.appledental.R
+
+class DoctorActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_doctor)
+    }
+}
