+++ app/src/main/java/app/kushagra/appledental/sections/doctor/DoctorVM.kt
@@ -1,0 +1,6 @@
+package app.kushagra.appledental.sections.doctor
+
+import app.kushagra.appledental.base.BaseViewModel
+
+class DoctorVM : BaseViewModel() {
+}
