+++ app/src/main/java/app/kushagra/appledental/sections/splash/SplashVM.kt
@@ -1,0 +1,8 @@
+package app.kushagra.appledental.sections.splash
+
+import app.kushagra.appledental.base.BaseViewModel
+
+class SplashVM : BaseViewModel() {
+
+
+}
