--- app/src/main/java/app/kushagra/takeoff/base/BaseViewModel.kt
@@ -1,21 +1,0 @@
-package app.kushagra.takeoff.base
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-abstract class BaseViewModel : ViewModel() {
-
-    val inProgress by lazy { MutableLiveData<Boolean>() }
-    val error by lazy { MutableLiveData<Throwable>() }
-
-    override fun onCleared() {
-        super.onCleared()
-        inProgress.value = false
-    }
-
-    fun onErrors(throwable: Throwable){
-        inProgress.value = false
-        error.value = throwable
-    }
-
-}
