--- app/src/main/java/app/kushagra/takeoff/base/CoordinatedFragment.kt
@@ -1,8 +1,0 @@
-package app.kushagra.takeoff.base
-
-import androidx.fragment.app.Fragment
-
-abstract class CoordinatedFragment : Fragment() {
-
-    abstract fun getTitle():String
-}
