--- app/src/main/java/app/kushagra/takeoff/extensions/EasyExtension.kt
@@ -1,2 +1,0 @@
-package app.kushagra.takeoff.extensions
-
