--- app/src/main/java/app/kushagra/takeoff/sections/MainActivity.kt
@@ -1,12 +1,0 @@
-package app.kushagra.takeoff.sections
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import app.kushagra.takeoff.R
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
