--- app/src/main/java/app/kushagra/takeoff/sections/doctor/DoctorActivity.kt
@@ -1,12 +1,0 @@
-package app.kushagra.takeoff.sections.doctor
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import app.kushagra.takeoff.R
-
-class DoctorActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_doctor)
-    }
-}
