--- app/src/main/java/app/kushagra/takeoff/sections/doctor/DoctorVM.kt
@@ -1,6 +1,0 @@
-package app.kushagra.takeoff.sections.doctor
-
-import app.kushagra.takeoff.base.BaseViewModel
-
-class DoctorVM : BaseViewModel() {
-}
