--- app/src/main/java/app/kushagra/takeoff/sections/splash/SplashVM.kt
@@ -1,9 +1,0 @@
-package app.kushagra.takeoff.sections.splash
-
-import androidx.lifecycle.ViewModel
-import app.kushagra.takeoff.base.BaseViewModel
-
-class SplashVM : BaseViewModel() {
-
-
-}
